// Chess3D/frontend/src/scenes/BasicScene/types.ts
import { ColorRepresentation, WebGLRenderer, Vector3 } from "three";
import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader.js";

export interface BasicSceneOptions {
    addGridHelper: boolean;
    lightHelpers?: boolean;
    cannonDebugger?: boolean;
}
export interface BasicSceneProps {
    renderer: WebGLRenderer;
    loader: GLTFLoader;
    options: BasicSceneOptions;
}
export interface LightOptions {
    color: ColorRepresentation;
    position: Vector3;
    intensity: number;
    lookAt?: Vector3;
    castShadow?: boolean;
}